requirements.txt
bitsandbytes
accelerate
xformers==0.0.29.post3
peft
trl
triton
cut_cross_entropy
unsloth_zoo
sentencepiece
protobuf
datasets
huggingface_hub
hf_transfer
unsloth
snac
librosa
soundfile
tensorboard
openai-whisper
PyAudio
